Mystic Doors review + Togglable SrvMessage-BossHP + Map-Event patch
Reviewed Mystic Doors.
Fixed several issues showing up on Duey in uncommon scenarios.
Fixed a concurrency issue with XMLDomMapleData.
Scheduled forward the "lock disposal" action within the source. Now, it's expected that, after a set while, no method should require usage of a disposed lock and, during that while, a supposed "disposed lock" is still available to run (although no new processes is expected to require use of these locks).
Fixed concurrency issues with player's current event instance, generating several inconsistencies when swiftly registering/unregistering from events.
Implemented a mutually exclusive approach for server message - Boss HPbar.
Fixed item-making Kage requiring lv71~80 ETC instead of the expected 81~90.
Removed the possibility to buy cosmetic coupons with mesos through the NPCs.
Sleepywood JQ's no longer gives cash items when they finish the quest repeatedly.
Added Duey trucks in several maps lacking it. Added NPC Duey in New Leaf City.
Fixed scripted quests not calculating QUEST_RATE (if applied) when rewarding experience and meso.

+++ src/net/server/audit/ThreadTracker.java
@@ -57,7 +57,7 @@ public class ThreadTracker {
     private final Map<Long, AtomicInteger> lockCount = new HashMap<>();
     private final Map<Long, MonitoredLockType> lockIds = new HashMap<>();
     private final Map<Long, Long> lockThreads = new HashMap<>();
-    private final Map<Long, Byte> lockUpdate = new HashMap<>();
+    private final Map<Long, Integer> lockUpdate = new HashMap<>();
     
     private final Map<MonitoredLockType, Map<Long, Integer>> locks = new HashMap<>();
     ScheduledFuture<?> threadTrackerSchedule;
@@ -169,8 +169,8 @@ public class ThreadTracker {
 
                     toRemove.clear();
 
-                    for(Entry<Long, Byte> it : lockUpdate.entrySet()) {
-                        byte val = (byte)(it.getValue() + 1);
+                    for(Entry<Long, Integer> it : lockUpdate.entrySet()) {
+                        int val = it.getValue() + 1;
 
                         if(val < 60) {
                             lockUpdate.put(it.getKey(), val);
@@ -202,7 +202,7 @@ public class ThreadTracker {
                         lockCount.put(lockOid, c);
                         lockIds.put(lockOid, lockId);
                         lockThreads.put(lockOid, tid);
-                        lockUpdate.put(lockOid, (byte) 0);
+                        lockUpdate.put(lockOid, 0);
                     }
                     c.incrementAndGet();
 
@@ -233,7 +233,7 @@ public class ThreadTracker {
                 else {
                     AtomicInteger c = lockCount.get(lockOid);
                     c.decrementAndGet();
-                    lockUpdate.put(lockOid, (byte) 0);
+                    lockUpdate.put(lockOid, 0);
 
                     List<MonitoredLockType> list = threadTracker.get(tid);
                     for(int i = list.size() - 1; i >= 0; i--) {
